python-server/requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Deprecated==1.2.7
Flask==1.1.1
Flask-Cors==3.0.8
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
PyGithub==1.47
PyJWT==1.7.1
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.1
